requirements.txt
aws-cdk-lib==2.122.0
constructs>=10.0.0,<11.0.0
aws_cdk.aws_appconfig_alpha==2.122.0a0
